change distance field to float

+++ database/migrations/20180416090102_createBikesTable.js
@@ -13,7 +13,7 @@ exports.up = (knex, Promise) => {
     table.varchar('brand', 255).notNullable().defaultTo('')
     table.varchar('model', 255).notNullable().defaultTo('')
     table.varchar('strava_gear_id', 60).notNullable().defaultTo('')
-    table.integer('distance').notNullable().defaultTo(0)
+    table.float('distance').notNullable().defaultTo(0)
     table.varchar('distance_unit', 20).notNullable()
     table.uuid('user_id').notNullable()
     table.foreign('user_id').references('id').inTable(usersTable).onDelete('cascade')
